checkers: move simplifiable condition checkers to SimplifiableIf
