+++ edulint/linting/checkers/simplifiable_if.py
@@ -1,5 +1,6 @@
 from astroid import nodes  # type: ignore
-from typing import TYPE_CHECKING, Optional, Tuple, Union, List
+from astroid.const import Context
+from typing import TYPE_CHECKING, Optional, Tuple, Union, List, Any, Dict, Set
 
 from pylint.checkers import BaseChecker  # type: ignore
 from pylint.checkers.utils import only_required_for_messages
@@ -7,7 +8,24 @@ from pylint.checkers.utils import only_required_for_messages
 if TYPE_CHECKING:
     from pylint.lint import PyLinter  # type: ignore
 
-from edulint.linting.checkers.utils import get_name, get_assigned_to, is_any_assign
+from edulint.linting.checkers.utils import (
+    get_name,
+    get_assigned_to,
+    is_any_assign,
+    is_pure_builtin,
+    get_const_value,
+)
+
+ExprRepresentation = str
+Comparison = str
+Value = Any
+NodeCmpValue = Optional[Tuple[nodes.NodeNG, Comparison, Value]]
+SameDirComparisons = Dict[
+    ExprRepresentation, Dict[Comparison, Tuple[List[ExprRepresentation], List[bool]]]
+]
+# When you take a key (1), then one key (cmp) from the corresponding dictionary and an element from the
+# first array from its corresponding value (2), you get a comparison ((1) (cmp) (2)). The List[bool]
+# indicates, whether the comparison in the original source code is in the same direction as in the dictionary.
 
 
 class SimplifiableIf(BaseChecker):  # type: ignore
@@ -67,6 +85,21 @@ class SimplifiableIf(BaseChecker):  # type: ignore
             "simplifiable-if-seq",
             "Emitted when the condition of an if statement can be merged "
             "with the next if's condition using logical operators.",
+        ),
+        "R6211": (
+            "'%s' can be replaced with '%s'",
+            "simplifiable-with-abs",
+            "Emitted when there is a problem like x < 4 and x > -4 and suggests abs(x) < 4.",
+        ),
+        "R6212": (
+            "'%s' can be replaced with '%s'",
+            "redundant-compare-in-condition",
+            "Emitted when there is a problem like x > 4 or x > 3 and suggests x > 3. (ie min{4, 3})",
+        ),
+        "R6213": (
+            "'%s' can be replaced with '%s'",
+            "redundant-compare-avoidable-with-max-min",
+            "Emitted when there is a problem like 'x > a and x > b' and suggests x > max(a, b).",
         ),
     }
 
@@ -313,6 +346,581 @@ class SimplifiableIf(BaseChecker):  # type: ignore
         if refactored is not None:
             new_cond, only_replaces = refactored
             self._simplifiable_if_message(node, then, new_cond, only_replaces)
+
+    SWITCHED_COMPARATOR = {
+        ">=": "<=",
+        ">": "<",
+        "<=": ">=",
+        "<": ">",
+    }
+
+    def _get_values_and_comparator(
+        self, node: nodes.Compare
+    ) -> Tuple[nodes.NodeNG, str, nodes.NodeNG]:
+        """
+        This method assumes node is a comparison between just two values.
+        """
+        left, (comp, right) = node.left, node.ops[0]
+        return left, comp, right
+
+    def _is_numeric_constant(self, node: nodes.NodeNG) -> bool:
+        return isinstance(node, nodes.Const) and (
+            isinstance(node.value, int) or isinstance(node.value, float)
+        )
+
+    def _is_number(self, node: nodes.NodeNG) -> bool:
+        return self._is_numeric_constant(node) or (
+            isinstance(node, nodes.UnaryOp)
+            and node.op in ["+", "-"]
+            and self._is_numeric_constant(node.operand)
+        )
+
+    def _get_node_comparator_const_value(self, node: nodes.NodeNG) -> NodeCmpValue:
+        """
+        Assumes that node is a child of Bool Op.
+        If the extraction was not successful returns None.
+        It is succesful if and only if it is an inequality between
+        a constant and a non-constant.
+
+        When this does not return None, the constant is last.
+        """
+        if not isinstance(node, nodes.Compare) or len(node.ops) >= 2:
+            return None
+
+        left, comp, right = self._get_values_and_comparator(node)
+
+        left_is_number = self._is_number(left)
+        right_is_number = self._is_number(right)
+
+        if (
+            (left_is_number and right_is_number)
+            or (not left_is_number and not right_is_number)
+            or comp not in self.SWITCHED_COMPARATOR
+        ):
+            return None
+
+        if left_is_number:
+            left, comp, right = right, self.SWITCHED_COMPARATOR[comp], left
+
+        return left, comp, get_const_value(right)
+
+    def _get_comparator_by_restriction(self, cmp1: str, cmp2: str, more_restrictive: bool) -> str:
+        if len(cmp1) > len(cmp2):
+            cmp1, cmp2 = cmp2, cmp1
+
+        return cmp1 if more_restrictive else cmp2
+
+    def _remove_redundant_compare(
+        self, cmp1: str, const1: Any, cmp2: str, const2: Any, is_and: bool
+    ) -> Tuple[str, int]:
+        """
+        Supposes that cmp1 and cmp2 are comparisons in the same direction (i.e. > and >=, ...)
+        and that const1 and const2 are both numric values.
+
+        <is_and> is True when there is <and> operation between the two comparisons and
+        is False when there is <or> between them.
+        """
+        if const1 == const2:
+            return [self._get_comparator_by_restriction(cmp1, cmp2, is_and), const1]
+
+        # make the const1 smaller than const2
+        if const1 > const2:
+            cmp1, const1, cmp2, const2 = cmp2, const2, cmp1, const1
+
+        if cmp1 == ">" or cmp1 == ">=":
+            if is_and:
+                return [cmp2, const2]
+            return [cmp1, const1]
+
+        if is_and:
+            return [cmp1, const1]
+
+        return [cmp2, const2]
+
+    def _add_abs_if_necessary(self, expr_string: str) -> str:
+        if expr_string.startswith("abs("):
+            return expr_string + " "
+
+        return "abs(" + expr_string + ") "
+
+    def _check_for_simplification_of_single_or(
+        self, const1: Any, cmp1: str, const2: Any, cmp2: str, expr_string: str
+    ) -> str:
+        """
+        Supposes that <expr_string> is a variable name or is of type f(x), where
+        f is pure builtin function.
+
+        And checks if an expression <<expr_string> <cmp1> <const1> or
+        <expr_string> <cmp2> <const2>> can be simplified.
+
+        Checks for problems like 'f(x) >= 4 or f(x) <= -4' and suggests to
+        replace them with 'abs(f(x)) >= 4'.
+
+        Returns the suggestion as a string.
+        """
+        new_expr = "True"
+
+        if cmp1 in {">=", ">"}:
+            if const1 > 0:
+                new_expr = self._add_abs_if_necessary(expr_string) + cmp1 + " " + str(const1)
+        else:
+            if const1 < 0:
+                new_expr = self._add_abs_if_necessary(expr_string) + cmp2 + " " + str(const2)
+
+        return new_expr
+
+    def _check_for_simplification_of_single_and(
+        self, const1: Any, cmp1: str, const2: Any, cmp2: str, expr_string: str
+    ) -> str:
+        """
+        Supposes that <expr_string> is a variable name or is of type f(x), where
+        f is pure builtin function. And also supposes that cmp1 and cmp2 are in
+        opposite directions.
+
+        And checks if an expression <<expr_string> <cmp1> <const1> and
+        <expr_string> <cmp2> <const2>> can be simplified.
+
+        Checks for problems like 'f(x) < 4 and f(x) > -4' and suggests to
+        replace them with 'abs(f(x)) < 4'.
+
+        Returns the suggestion as a string.
+        """
+        new_expr = "False"
+
+        if cmp1 in {">=", ">"}:
+            if const1 < 0:
+                new_expr = self._add_abs_if_necessary(expr_string) + cmp2 + " " + str(const2)
+        else:
+            if const1 > 0:
+                new_expr = self._add_abs_if_necessary(expr_string) + cmp1 + " " + str(const1)
+
+        return new_expr
+
+    def _check_if_always_true_or_false(
+        self, const1: Any, cmp1: str, const2: Any, cmp2: str, expr_string: str, op: str
+    ) -> str:
+        if const1 == const2:
+            if len(cmp1) == 1 and len(cmp2) == 1 and op == "or":
+                return expr_string + " != " + str(const1)
+
+            if len(cmp1) == 2 and len(cmp2) == 2 and op == "and":
+                return expr_string + " == " + str(const1)
+
+            if op == "or":
+                return "True"
+
+            return "False"
+
+        if op == "and" and (
+            (const1 < const2 and cmp1[0] == "<") or (const1 > const2 and cmp2[0] == "<")
+        ):
+            return "False"
+
+        if op == "or" and (
+            (const1 < const2 and cmp1[0] == ">") or (const1 > const2 and cmp2[0] == ">")
+        ):
+            return "True"
+
+        return ""
+
+    def _get_group_as_string(
+        self, expr_string: str, group: List[Tuple[Comparison, Value]], op: str
+    ) -> str:
+        result = []
+
+        for i in range(len(group)):
+            cmp, val = group[i]
+            result.extend([expr_string, cmp, str(val)])
+            if i < len(group) - 1:
+                result.append(op)
+
+        return " ".join(result)
+
+    def _suggestion_for_group_of_boolops(
+        self, expr_string: str, group: List[Tuple[Comparison, Value]], node: nodes.BoolOp
+    ) -> None:
+        # We'll use commutativity of boolean operations and group the comparisons in same direction
+        # (< with <= and > with >=) and simplify the grouped parts and then we can simplify the rest.
+        cmp_less, val1 = None, None
+        cmp_greater, val2 = None, None
+        changed = False
+
+        for cmp, val in group:
+            if cmp[0] == "<":
+                if cmp_less is None:
+                    cmp_less, val1 = cmp, val
+                    continue
+                changed = True
+                cmp_less, val1 = self._remove_redundant_compare(
+                    cmp_less, val1, cmp, val, node.op == "and"
+                )
+            else:
+                if cmp_greater is None:
+                    cmp_greater, val2 = cmp, val
+                    continue
+                changed = True
+                cmp_greater, val2 = self._remove_redundant_compare(
+                    cmp_greater, val2, cmp, val, node.op == "and"
+                )
+
+        group_string = self._get_group_as_string(expr_string, group, node.op)
+
+        if cmp_less is None:
+            self.add_message(
+                "redundant-compare-in-condition",
+                node=node,
+                args=(group_string, " ".join([expr_string, cmp_greater, str(val2)])),
+            )
+            return
+
+        if cmp_greater is None:
+            self.add_message(
+                "redundant-compare-in-condition",
+                node=node,
+                args=(group_string, " ".join([expr_string, cmp_less, str(val1)])),
+            )
+            return
+
+        new_expr = self._check_if_always_true_or_false(
+            val1, cmp_less, val2, cmp_greater, expr_string, node.op
+        )
+
+        if len(new_expr) != 0:
+            self.add_message(
+                "redundant-compare-in-condition",
+                node=node,
+                args=(group_string, new_expr),
+            )
+            return
+
+        if val1 != -val2 or cmp_less != self.SWITCHED_COMPARATOR[cmp_greater]:
+            if changed:
+                self.add_message(
+                    "redundant-compare-in-condition",
+                    node=node,
+                    args=(
+                        group_string,
+                        " ".join(
+                            [
+                                expr_string,
+                                cmp_less,
+                                str(val1),
+                                node.op,
+                                expr_string,
+                                cmp_greater,
+                                str(val2),
+                            ]
+                        ),
+                    ),
+                )
+            return
+
+        if node.op == "and":
+            new_expr = self._check_for_simplification_of_single_and(
+                val1, cmp_less, val2, cmp_greater, expr_string
+            )
+        else:
+            new_expr = self._check_for_simplification_of_single_or(
+                val1, cmp_less, val2, cmp_greater, expr_string
+            )
+
+        self.add_message(
+            "simplifiable-with-abs",
+            node=node,
+            args=(group_string, new_expr),
+        )
+
+    def _is_pure_node(self, node: nodes.NodeNG):
+        """
+        Note: This method does not check children of the node, just the node itself.
+        """
+        return (
+            isinstance(node, nodes.Name)
+            or isinstance(node, nodes.Const)
+            or isinstance(node, nodes.BinOp)
+            or isinstance(node, nodes.BoolOp)
+            or isinstance(node, nodes.UnaryOp)
+            or isinstance(node, nodes.Compare)
+            or (isinstance(node, nodes.Subscript) and node.ctx == Context.Load)
+            or isinstance(node, nodes.Attribute)
+            or (
+                isinstance(node, nodes.Call)
+                and len(node.keywords) == 0
+                and len(node.kwargs) == 0
+                and len(node.starargs) == 0
+                and is_pure_builtin(node.func)
+            )
+        )
+
+    def _is_pure_expression(self, node: nodes.NodeNG) -> bool:
+        if not self._is_pure_node(node):
+            return False
+
+        children = node.get_children()
+        if isinstance(node, nodes.Call):
+            # the first child of Call is always the function itself and we know it is pure by now.
+            next(children)
+
+        for child in children:
+            if not self._is_pure_expression(child):
+                return False
+
+        return True
+
+    def _group_and_check_by_representation(
+        self, comparison_operands: List[NodeCmpValue], node: nodes.BoolOp
+    ) -> None:
+        already_checked: Set[int] = set()
+        current_group: List[Tuple[Comparison, Value]] = []
+
+        for i in range(len(comparison_operands)):
+            if i in already_checked:
+                continue
+
+            expr1, cmp1, const1 = comparison_operands[i]
+            current_group = [(cmp1, const1)]
+            for j in range(i + 1, len(comparison_operands)):
+                if j in already_checked:
+                    continue
+
+                expr2, cmp2, const2 = comparison_operands[j]
+                # I could have done something more, but thanks to how as_string returns same string
+                # even if there are some additional parentheses or spaces this seems good enough.
+                # Could be improved with checking the actual structure and types of expr1 and expr2,
+                # but with the knowledge that all the operands are pure and are from the same bool
+                # expression we can safely just check using as_string().
+                if expr1.as_string() == expr2.as_string():
+                    current_group.append((cmp2, const2))
+                    already_checked.add(j)
+
+            if len(current_group) < 2:
+                continue
+
+            self._suggestion_for_group_of_boolops(expr1.as_string(), current_group, node)
+
+    def _update_comparisons(
+        self,
+        comparisons: SameDirComparisons,
+        expr1: str,
+        comparator: str,
+        expr2: str,
+        comparison_switched: bool,
+    ):
+        if expr1 in comparisons:
+            if comparator in comparisons[expr1]:
+                comparisons[expr1][comparator][0].append(expr2)
+                comparisons[expr1][comparator][1].append(comparison_switched)
+            else:
+                comparisons[expr1][comparator] = ([expr2], [comparison_switched])
+        else:
+            comparisons[expr1] = {comparator: ([expr2], [comparison_switched])}
+
+    def _switch_if_necessary(
+        self, expr1: str, cmp: str, expr2: str, was_switched: bool
+    ) -> List[str]:
+        if was_switched:
+            return [expr2, self.SWITCHED_COMPARATOR[cmp], expr1]
+
+        return [expr1, cmp, expr2]
+
+    def _get_original_group_as_string(
+        self,
+        expr_string: str,
+        cmp: str,
+        group: List[ExprRepresentation],
+        was_switched: List[bool],
+        compared_with_num: Optional[Tuple[str, bool]],
+        op: str,
+    ) -> str:
+        result = []
+
+        for i in range(len(group)):
+            result.extend(self._switch_if_necessary(expr_string, cmp, group[i], was_switched[i]))
+
+            if i < len(group) - 1:
+                result.append(op)
+
+        if compared_with_num is not None:
+            result.append(op)
+            result.extend(
+                self._switch_if_necessary(
+                    expr_string,
+                    cmp,
+                    compared_with_num[0],
+                    compared_with_num[1],
+                )
+            )
+
+        return " ".join(result)
+
+    def _put_into_max_or_min(
+        self,
+        function: str,
+        expression: str,
+        comparison: str,
+        expressions: List[str],
+        compared_with_num: Optional[Tuple[str, bool]],
+    ) -> str:
+        """
+        the <function> should contain either "max(" or "min("
+        """
+        result: List[str] = [expression + " " + comparison + " " + function]
+
+        for i in range(len(expressions)):
+            if expressions[i].startswith(function):
+                result.append(expressions[i][4:-1])
+            else:
+                result.append(expressions[i])
+
+            if i < len(expressions) - 1:
+                result.append(", ")
+
+        if compared_with_num is not None:
+            result.append(", ")
+            result.append(compared_with_num[0])
+
+        result.append(")")
+        return "".join(result)
+
+    def _make_suggestion_for_using_max_min_if_possible(
+        self,
+        comparisons_with_numbers: SameDirComparisons,
+        comparisons: SameDirComparisons,
+        node: nodes.BoolOp,
+    ) -> None:
+        """
+        Makes suggestion only for the comparisons that will maximize the number of
+        comparisons put together with max (ie. in here: "x > a and b < x and b < 1
+        and a > b" would suggest "x > a and b < min(x, 1, a)" instead of
+        "x > max(a, b) and b < 1 and b < a"). And makes only one suggestion (after
+        finding the suggestion as described earlier the method just returns).
+        """
+        expression: Optional[str] = None
+        comparison: Optional[str] = None
+        expressions: Optional[List[str]] = None
+        was_switched: Optional[List[bool]] = None
+        compared_with_num: Optional[Tuple[str, bool]] = None
+
+        for expr, compared_with in comparisons.items():
+            for cmp, exprs in compared_with.items():
+                if len(exprs[0]) < 2:
+                    continue
+
+                comparison_count = len(exprs[0])
+                compared_with_1_number = False
+                # note: when it's compared with multiple numbers the redundant compare in condition
+                # or simplifiable-with-abs could make a suggestion and I don't want to make
+                # conflicting suggestions.
+
+                if (
+                    expr in comparisons_with_numbers
+                    and len(comparisons_with_numbers[expr]) == 1
+                    and cmp in comparisons_with_numbers[expr]
+                    and len(comparisons_with_numbers[expr][cmp][0]) == 1
+                ):
+                    comparison_count += 1
+                    compared_with_1_number = True
+
+                if expression is None or len(expressions) < comparison_count:
+                    expression = expr
+                    comparison = cmp
+                    expressions, was_switched = exprs
+
+                    if compared_with_1_number:
+                        compared_with_num = (
+                            comparisons_with_numbers[expr][cmp][0][0],
+                            comparisons_with_numbers[expr][cmp][1][0],
+                        )
+                    else:
+                        compared_with_num = None
+
+        if expression is None:
+            return
+
+        group_string = self._get_original_group_as_string(
+            expression, comparison, expressions, was_switched, compared_with_num, node.op
+        )
+
+        func = (
+            "max("
+            if (
+                (node.op == "and" and comparison[0] == ">")
+                or (node.op == "or" and comparison[0] == "<")
+            )
+            else "min("
+        )
+
+        self.add_message(
+            "redundant-compare-avoidable-with-max-min",
+            node=node,
+            args=(
+                group_string,
+                self._put_into_max_or_min(
+                    func, expression, comparison, expressions, compared_with_num
+                ),
+            ),
+        )
+
+    def _is_comparison_not_between_numbers(self, node: nodes.NodeNG) -> bool:
+        return (
+            isinstance(node, nodes.Compare)
+            and len(node.ops) == 1
+            and node.ops[0][0] in self.SWITCHED_COMPARATOR
+            and not self._is_number(node.left)
+            and not self._is_number(node.ops[0][1])
+        )
+
+    def _check_for_simplification_of_boolop(self, node: nodes.BoolOp) -> None:
+        if node.op is None:
+            return
+
+        comparison_operands: List[NodeCmpValue] = []
+        comparisons_with_numbers: SameDirComparisons = {}
+        comparisons: SameDirComparisons = {}
+
+        for value in node.values:
+            if not self._is_pure_expression(value):
+                return
+
+            operand = self._get_node_comparator_const_value(value)
+
+            if operand is not None:
+                comparison_operands.append(operand)
+                self._update_comparisons(
+                    comparisons_with_numbers,
+                    operand[0].as_string(),
+                    operand[1],
+                    str(operand[2]),
+                    value.ops[0][0] != operand[1],
+                )
+            elif self._is_comparison_not_between_numbers(value):
+                self._update_comparisons(
+                    comparisons,
+                    value.left.as_string(),
+                    value.ops[0][0],
+                    value.ops[0][1].as_string(),
+                    False,
+                )
+                self._update_comparisons(
+                    comparisons,
+                    value.ops[0][1].as_string(),
+                    self.SWITCHED_COMPARATOR[value.ops[0][0]],
+                    value.left.as_string(),
+                    True,
+                )
+
+        self._group_and_check_by_representation(comparison_operands, node)
+        self._make_suggestion_for_using_max_min_if_possible(
+            comparisons_with_numbers, comparisons, node
+        )
+
+    @only_required_for_messages(
+        "simplifiable-with-abs",
+        "redundant-compare-in-condition",
+        "redundant-compare-avoidable-with-max-min",
+    )
+    def visit_boolop(self, node: nodes.BoolOp) -> None:
+        self._check_for_simplification_of_boolop(node)
 
 
 def register(linter: "PyLinter") -> None:
